refactor population flow
