+++ app/app.go
@@ -6,9 +6,9 @@ import (
 	"github.com/go-kid/ioc/configure"
 	"github.com/go-kid/ioc/definition"
 	"github.com/go-kid/ioc/factory"
-	"github.com/go-kid/ioc/factory/processors/definition_registry_post_processors"
-	"github.com/go-kid/ioc/factory/processors/instantiation_aware_component_post_processors"
 	"github.com/go-kid/ioc/factory/support"
+	"github.com/go-kid/ioc/processors/definition_registry_post_processors"
+	"github.com/go-kid/ioc/processors/instantiation_aware_component_post_processors"
 	"github.com/go-kid/ioc/syslog"
 	"github.com/go-kid/ioc/util/sort2"
 	"sync"
