+++ factory/post_processor_registration_delegate.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"github.com/go-kid/ioc/component_definition"
 	"github.com/go-kid/ioc/definition"
-	"github.com/go-kid/ioc/factory/processors"
+	"github.com/go-kid/ioc/processors"
 	"github.com/go-kid/ioc/syslog"
 	"github.com/go-kid/ioc/util/reflectx"
 	"github.com/go-kid/ioc/util/sort2"
