+++ processors/default_component_post_processor.go
