+++ processors/default_instantiation_aware_component_post_processor.go
